Move to executable/class_definition_evaluation.rs
